+++ application/src/components/index.js
@@ -1,7 +1,7 @@
 import Login from "./login";
 import Main from "./main/main";
 import Nav from "./nav/nav";
-import OrderFormHook from "./order-form-hook/order-form";
+import OrderForm from "./order-form/OrderForm";
 import Signup from "./signup";
 import Template from "./common/template";
 import ViewOrders from "./view-orders/viewOrders";
@@ -9,7 +9,7 @@ import ViewOrdersHook from "./view-orders-hook/viewOrders";
 
 export {
   Login,
-  OrderFormHook,
+  OrderForm,
   Main,
   Nav,
   Signup,
